tests: rewrite the new test according to latest rebase
